+++ routes/router.js
@@ -70,8 +70,8 @@ router.route('/chat/addFriend').post(protectPath, addFriendToChat) //NOT TESTED!
 router.route('/createOrRetrieveChat').post(protectPath, checkMutualChat)
 router.route('/chat').post(protectPath, createChat)
 router.route('/chat/leaveChat').post(protectPath, leaveChat)
-router.route('/comment').post(protectPath, createComment)
 router.route('/chats').get(protectPath, getUsersChats)
+router.route('/comment').post(protectPath, createComment)
 router.route('/friends').get(protectPath, getFriends)
 router.route('/friend/:friendName').post(protectPath, makeFriendRequest).delete(protectPath, deleteFriend)
 router.route('/acceptFriendRequest/:requestCreatorName').post(protectPath, acceptFriendRequest)
